update 0000-01-02-shahalz.md
In Lemmatization root word is called Lemma. A lemma (plural lemmas or lemmata) is the canonical form, dictionary form, or citation form of a set of words. For example, runs, running, ran are all forms of the word run, therefore run is the lemma of all these words. Operation Mincemeat was a successful British deception operation of the Second World War. In 1943, the body of Glyndwr Michael was dressed as an officer of the Royal Marines, and items were placed on him identifying him as the fictitious Captain William Martin. Correspondence between two British generals suggesting that the Allies planned to invade Greece and Sardinia, with merely a feint at Sicily, was also planted. The body was taken by the submarine HMS Seraph to the southern coast of Spain and released close to shore, where it was picked up the following morning by a Spanish fisherman. The nominally neutral Spanish government shared copies of the documents with German military intelligence, before returning them to the British. Ultra decrypts of German messages showed that the Germans fell for the ruse. Reinforcements were shifted to Greece and Sardinia before and during the invasion of Sicily; Sicily received none. The body was buried with military honours in Spain (grave pictured). (Full article...)